[Blocks] Cleaning + Made background with blocks
Called the gridsettings seperatly in the constructor AND made these variables inside the class. So that it is easier to call inside the methods. Also made the background with a 2d-color array, where when the block stops on the bottom, this shape will get displayed on the background instead of disappearing.